fix: bugfix and balancing
- background music is now music not sound
- muni auto refill
- more than 1 customer spawn random
- money calculation fix
- PlayerStat showed hp/hp now hp/maxHp

- the waiting time by order is harder by higher level

+++ Assets/Scripts/ButtonHandler.cs
@@ -17,7 +17,7 @@ public class ButtonHandler : MonoBehaviour
 
     public void Upgrade(int upgradenr)
     {
-        int money = GameManager.Instance.m_Money;
+        int money = (int)GameManager.Instance.m_Money;
         int diamands = GameManager.Instance.m_Diamonds;
     }
